refactor: sql-tools readers
