+++ server/src/sql-tools/from-database/readers/type.ts
@@ -1,0 +1,3 @@
+import { DatabaseClient, DatabaseSchema } from 'src/sql-tools/types';
+
+export type DatabaseReader = (schema: DatabaseSchema, db: DatabaseClient) => Promise<void>;
